Fixed inst. of commands, added inst. for datatype against ftvar.
Instantiation of commands was not being performed at all for the
computation pattern matching rule and was being performed incorrectly
for the command inference rule. The problem manifested itself when the
typechecker tried to unify an interface's rigid type variable parameter
against the actual rigid provided in the effect set specification. These
have both been resolved to do The Right Thing(tm).

Checking a datatype against a flexible type variable was not
performing instantiation of the datatype's parameters as it
should and as it does now.

The state example works typechecks and evaluates to the expected
result.