Updated typechecker so print_map.fk example typechecks.
Require to handle construction of shapes of types when checking
against flexibles. For example, creating a flexible type variable
of the returner type shape when typechecking checkable values.

The change required a flexible type variable to appear in the
effects sets of types and the ambient effects within the environment.
The solution is to allow a flexible to be unified with a list of source
types and this structure will be unboxed when performing other parts of
typechecking.

I'm not particularly happy with this solution since it really creates
a list of effects inside the outer list, but it seemed to be the quickest
(albeit perhaps also the dirtiest!) way to get around the unification
issues I have been having while trying to get print_map.fk to typecheck.