Added does and sub-effecting calculations.
This has broken some programs but the errors would suggest it is
catching actual type errors rather than being incorrect.

Unification has to be performed in some cases -- when replacing
the effect variable with the ambient effects the ambient effects
may contain interfaces which are specified by a handler's signature.
The delimited continuations example highlighted the need for this.

So unification is performed when making a list unique (i.e. replacing
the effect variable with the ambient effects) and when testing the
equality of two effect sets (sub/<= relaton).