Fixed pattern matching bug and added corresponding test case.
The pattern matching code implemented a strict uniqueness condition
which obilerated the environment if a duplicate pattern was detected.
This behaviour should only be applied to a single set of patterns and
not across multiple clauses.

Moreover, pattern matching does not need to access to the global
environment since it is only queries the environment in order to
enforce this local uniqueness property. Only the body of the
clause needs the global environment (extended with the local patterns)
and so a simple merge operation is performed.