Added a fix for the bind-eff.fk test case (see notes).
The current version of main will now typecheck. This error was
introduced during a previous commit swapping the order of
effect variable instantiation and uniqueness enforcing of the
effect set. Adding a case for flexible tvars resolved this issue.

However, the other three cases which are commented out still fail
to typecheck. The issue in those cases are different and further
investigation is required. These cases will be separated out into
their own regression test files.