Added evaluation of command patterns.
Rewriting the evaluation code has simplified significantly the
definitions of matching on commands and values. The environment is
updated in only one place and redundant bindings (for constructors
and commands) have been removed.

The next stage is to consider how open effect sets are handled and
how forwarding is treated.